src/user_interface.py
```python
import tkinter as tk
from tkinter import filedialog
from dance_styles import selectDanceStyle
from music_integration import selectMusicTrack
from motion_analysis import analyzeMotionData
from choreography_generation import generateChoreography
from customization import customizeChoreography
from collaboration_sharing import shareChoreography
from tutorial_resources import accessTutorialResource
from performance_assessment import assessPerformance
from monetization import updateSubscription, makeInAppPurchase
from partnerships import updatePartnershipStatus

class UserInterface:
    def __init__(self, root):
        self.root = root
        self.root.title("AI Dance Choreographer")
        self.create_widgets()

    def create_widgets(self):
        self.dance_style_button = tk.Button(self.root, text="Select Dance Style", command=self.select_dance_style)
        self.dance_style_button.pack()

        self.music_track_button = tk.Button(self.root, text="Select Music Track", command=self.select_music_track)
        self.music_track_button.pack()

        self.motion_analysis_button = tk.Button(self.root, text="Analyze Motion Data", command=self.analyze_motion_data)
        self.motion_analysis_button.pack()

        self.generate_choreography_button = tk.Button(self.root, text="Generate Choreography", command=self.generate_choreography)
        self.generate_choreography_button.pack()

        self.customize_choreography_button = tk.Button(self.root, text="Customize Choreography", command=self.customize_choreography)
        self.customize_choreography_button.pack()

        self.share_choreography_button = tk.Button(self.root, text="Share Choreography", command=self.share_choreography)
        self.share_choreography_button.pack()

        self.access_tutorial_resource_button = tk.Button(self.root, text="Access Tutorial Resources", command=self.access_tutorial_resource)
        self.access_tutorial_resource_button.pack()

        self.assess_performance_button = tk.Button(self.root, text="Assess Performance", command=self.assess_performance)
        self.assess_performance_button.pack()

        self.update_subscription_button = tk.Button(self.root, text="Update Subscription", command=self.update_subscription)
        self.update_subscription_button.pack()

        self.make_in_app_purchase_button = tk.Button(self.root, text="Make In-App Purchase", command=self.make_in_app_purchase)
        self.make_in_app_purchase_button.pack()

        self.update_partnership_status_button = tk.Button(self.root, text="Update Partnership Status", command=self.update_partnership_status)
        self.update_partnership_status_button.pack()

    def select_dance_style(self):
        selectDanceStyle()

    def select_music_track(self):
        filename = filedialog.askopenfilename()
        selectMusicTrack(filename)

    def analyze_motion_data(self):
        analyzeMotionData()

    def generate_choreography(self):
        generateChoreography()

    def customize_choreography(self):
        customizeChoreography()

    def share_choreography(self):
        shareChoreography()

    def access_tutorial_resource(self):
        accessTutorialResource()

    def assess_performance(self):
        assessPerformance()

    def update_subscription(self):
        updateSubscription()

    def make_in_app_purchase(self):
        makeInAppPurchase()

    def update_partnership_status(self):
        updatePartnershipStatus()

if __name__ == "__main__":
    root = tk.Tk()
    app = UserInterface(root)
    root.mainloop()
```